+++ common-core/src/main/java/cn/featherfly/common/bean/BeanPropertyValue.java
@@ -52,4 +52,12 @@ public class BeanPropertyValue<T> {
         return value;
     }
 
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public String toString() {
+        return "[beanProperty=" + beanProperty + ", value=" + value + "]";
+    }
+
 }