+++ common-core/change-log.md
@@ -1,6 +1,7 @@
-# 1.8
+# 1.8.33 2022-03-17
     1.storage加入boolean exists(ID id)方法
     2.BeanPropertyValue实现自定义toString()方法
+    3.BeanProperty加入List<Class<?>> getGenericTypes()方法
 
 # 1.8.32 2021-12-05
     1. 加入BeanPropertyValue